+++ packages/block-library/src/index.js
@@ -93,6 +93,8 @@ import * as pattern from './pattern';
 import * as pageList from './page-list';
 import * as pageListItem from './page-list-item';
 import * as paragraph from './paragraph';
+import * as playlist from './playlist';
+import * as playlistTrack from './playlist-track';
 import * as postAuthor from './post-author';
 import * as postAuthorName from './post-author-name';
 import * as postAuthorBiography from './post-author-biography';
@@ -291,6 +293,11 @@ const getAllBlocks = () => {
 
 	if ( window?.__experimentalNavigationOverlays ) {
 		blocks.push( navigationOverlayClose );
+	}
+
+	if ( window?.__experimentalEnableBlockExperiments ) {
+		blocks.push( playlist );
+		blocks.push( playlistTrack );
 	}
 
 	// When in a WordPress context, conditionally